fix(routers): cast queue_depth sort to int8, not int4 (phaze-5yyh)
/admin/agents' queue-sort ORDER BY cast the agent-supplied JSONB queue_depth to
SQLAlchemy Integer (Postgres int4, max 2147483647), but queue_depth is committed
as an int8 domain on both the heartbeat writer (QUEUE_DEPTH_MAX = 10**12) and the
storage side (_LANE_MERGE_SQL's ::bigint SUM). Any heartbeat above INT32_MAX --
a valid value on that domain -- made the whole statement raise DataError, 500ing
the page and, once queue-sorted, its own never-halting 5s self-poll for every
operator until the offending row was deleted. Cast to BigInteger so the read
side matches the bound the write side already commits to.

+++ src/phaze/routers/admin_agents.py
@@ -41,7 +41,7 @@ from typing import Annotated
 from fastapi import APIRouter, Depends, Query, Request
 from fastapi.responses import HTMLResponse
 from fastapi.templating import Jinja2Templates
-from sqlalchemy import DateTime, Integer, cast, func, literal, select
+from sqlalchemy import BigInteger, DateTime, cast, func, literal, select
 from sqlalchemy.ext.asyncio import AsyncSession  # noqa: TC002 — FastAPI needs runtime import to resolve Annotated[AsyncSession, Depends(...)]
 
 from phaze.config import get_settings
@@ -74,7 +74,15 @@ _LAST_SEEN_ORDER = func.coalesce(Agent.last_seen_at, cast(literal("-infinity"), 
 # Queue depth lives in the ``last_status`` JSONB blob and is absent for agents that never reported one
 # (the template renders those as "—"). Folded to -1 so they sort BELOW every real depth under ASC and
 # above nothing under DESC, rather than scattering on NULL ordering rules.
-_QUEUE_DEPTH_ORDER = func.coalesce(cast(Agent.last_status["queue_depth"].astext, Integer), -1)
+#
+# phaze-5yyh: cast to BigInteger (int8), NOT Integer (int4). ``queue_depth`` is committed as an int8
+# domain on both the write side (schemas/agent_heartbeat.py's ``QUEUE_DEPTH_MAX = 10**12``) and the
+# storage side (``_LANE_MERGE_SQL``'s ``SUM(... ::bigint)``), so a heartbeat above INT32_MAX (a valid
+# value on that domain) is a legitimate JSONB value this ORDER BY must be able to sort, not an error.
+# An int4 cast made Postgres raise DataError ("value ... out of range for type integer") on the whole
+# statement -- 500ing this table (and, once queue-sorted, its own never-halting 5s self-poll) for
+# every operator until the offending row was deleted.
+_QUEUE_DEPTH_ORDER = func.coalesce(cast(Agent.last_status["queue_depth"].astext, BigInteger), -1)
 
 AGENTS_SORT = SortContract(
     endpoint="/admin/agents/_table",
